+++ src/main/java/com/github/nylle/javafixture/extensions/testcases/ReflectedTestCase.java
@@ -1,4 +1,4 @@
-package com.github.nylle.javafixture.parameterized;
+package com.github.nylle.javafixture.extensions.testcases;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
