+++ src/main/java/com/github/nylle/javafixture/extensions/testcases/TestCaseException.java
@@ -1,4 +1,4 @@
-package com.github.nylle.javafixture.parameterized;
+package com.github.nylle.javafixture.extensions.testcases;
 
 public class TestCaseException extends RuntimeException {
