+++ src/main/java/com/github/nylle/javafixture/extensions/testcases/TestWithCases.java
@@ -1,4 +1,7 @@
-package com.github.nylle.javafixture.extension;
+package com.github.nylle.javafixture.extensions.testcases;
+
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ArgumentsSource;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -6,13 +9,10 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ArgumentsSource;
-
 @Documented
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-@ArgumentsSource(JavaFixtureProvider.class)
+@ArgumentsSource(TestCasesProvider.class)
 @ParameterizedTest
-public @interface TestWithFixture {
+public @interface TestWithCases {
 }
