+++ src/test/java/com/github/nylle/javafixture/SpecimenFactoryTest.java
@@ -1,8 +1,8 @@
 package com.github.nylle.javafixture;
 
+import com.github.nylle.javafixture.extensions.testcases.TestCase;
+import com.github.nylle.javafixture.extensions.testcases.TestWithCases;
 import com.github.nylle.javafixture.generic.FixtureType;
-import com.github.nylle.javafixture.parameterized.TestCase;
-import com.github.nylle.javafixture.parameterized.TestWithCases;
 import com.github.nylle.javafixture.specimen.ArraySpecimen;
 import com.github.nylle.javafixture.specimen.CollectionSpecimen;
 import com.github.nylle.javafixture.specimen.EnumSpecimen;
