+++ src/test/java/com/github/nylle/javafixture/extensions/fixture/JavaFixtureMethodExtensionTest.java
@@ -1,4 +1,4 @@
-package com.github.nylle.javafixture.extension;
+package com.github.nylle.javafixture.extensions.fixture;
 
 import com.github.nylle.javafixture.testobjects.example.Contract;
 import org.junit.jupiter.api.Test;
