+++ src/test/java/com/github/nylle/javafixture/extensions/fixture/TestWithFixtureTest.java
@@ -1,4 +1,4 @@
-package com.github.nylle.javafixture.extension;
+package com.github.nylle.javafixture.extensions.fixture;
 
 import com.github.nylle.javafixture.testobjects.example.Contract;
