+++ src/test/java/com/github/nylle/javafixture/extensions/testcases/TestCaseTest.java
@@ -1,6 +1,4 @@
-package com.github.nylle.javafixture.parameterized;
-
-import static org.assertj.core.api.Assertions.assertThat;
+package com.github.nylle.javafixture.extensions.testcases;
 
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -10,6 +8,8 @@ import java.util.NavigableSet;
 import java.util.Set;
 import java.util.SortedSet;
 import java.util.TreeSet;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 class TestCaseTest {
